+++ RELEASES.md
@@ -14,6 +14,8 @@ This new release adds support for sparse cost matrices and a new lazy EMD solver
 - Add support for sparse cost matrices in EMD solver (PR #778, Issue #397)
 - Added UOT1D with Frank-Wolfe in `ot.unbalanced.uot_1d` (PR #765)
 - Add Sliced UOT and Unbalanced Sliced OT in `ot/unbalanced/_sliced.py` (PR #765)
+- Wrapper for barycenter solvers with free support `ot.solvers.bary_free_support` (PR #730)
+
 
 #### Closed issues